requirements.txt
django
djangorestframework
dj-database-url
djangorestframework-simplejwt
django-filter
Pillow
gunicorn
